Add float32 decompress kernels and dtype-aware DF.loop
- unpack.cu: float32 counterparts of decompress_and_transpose and
  decompress_and_fill (d_t_kernel_f32 / decompress_kernel_f32) so an
  fp32 CDERI block can be unpacked without silent garbage (the double
  kernels reinterpret fp32 buffers)
- fill_symmetric dispatches on a.dtype; fp32 input with float64 out
  raises instead of corrupting
- DF.loop allocates the unpack workspace with the cderi dtype so a
  user-provided fp32 cderi (e.g. loaded from an fp32 checkpoint) works

An experiment storing fp32 CDERI at build time (fp32 GEMM for j3c @
aux_coef) was reverted: the 3080-long fp32 accumulation leaves ~5e-4
relative error in the CDERI which shifts the SCF fixed point (~3.8 Eh
after non-convergence at loose conv). The cderi tensor therefore stays
float64; fp32 speedup comes from contractions only.

Tamoxifen def2-SVP: fp64 5.22s / auto 3.63s (1.44x) / fp32@conv1e-5
1.76s (2.96x, err 1.7e-4)

+++ gpu4pyscf/df/df.py
@@ -26,6 +26,7 @@ from gpu4pyscf.lib.cupy_helper import (
 from gpu4pyscf.lib import logger
 from gpu4pyscf.lib import utils
 from gpu4pyscf.lib import multi_gpu
+from gpu4pyscf.lib import precision
 from gpu4pyscf.df import int3c2e_bdiv
 from gpu4pyscf.df import df_jk
 from gpu4pyscf.gto.mole import SortedMole
@@ -236,7 +237,7 @@ class DF(lib.StreamObject):
         on_gpu = isinstance(cderi_sparse, cp.ndarray)
 
         if unpack:
-            work = cp.zeros((nao, nao, blksize))
+            work = cp.zeros((nao, nao, blksize), dtype=cderi_sparse.dtype)
 
         pair_idx = cp.asarray(self._cderi_idx[0], dtype=np.int32)
         npairs = len(pair_idx)
@@ -251,7 +252,7 @@ class DF(lib.StreamObject):
                     yield out.transpose(2,0,1), cderi_blk
 
         else:
-            buf = cp.empty((blksize, npairs))
+            buf = cp.empty((blksize, npairs), dtype=cderi_sparse.dtype)
             buf_prefetch = cp.empty_like(buf)
 
             comput_stream = cp.cuda.get_current_stream()
